switched to long types for stride and indices, this fixed int overflowing problems which caused crash when writing to high-res volumes

+++ houdini/ocl/include/vft_utils.h
@@ -9,7 +9,7 @@
 
 
 // export float attrib
-static void vstore1(float dataIn, int i, global float* dataOut)
+static void vstore1(float dataIn, long i, global float* dataOut)
 {
     dataOut[i] = dataIn;
 }
